+++ node-api/src/main/kotlin/net/corda/nodeapi/internal/cryptoservice/CryptoService.kt
@@ -1,4 +1,4 @@
-package net.corda.node.internal.cryptoservice
+package net.corda.nodeapi.internal.cryptoservice
 
 import net.corda.core.DoNotImplement
 import org.bouncycastle.operator.ContentSigner
