+++ node-api/src/test/kotlin/net/corda/nodeapi/internal/crypto/AliasPrivateKeyTest.kt
@@ -1,6 +1,6 @@
 package net.corda.nodeapi.internal.crypto
 
-import net.corda.core.crypto.AliasPrivateKey
+import net.corda.core.crypto.internal.AliasPrivateKey
 import net.corda.testing.internal.stubs.CertificateStoreStubs
 import org.junit.Rule
 import org.junit.Test
