+++ node/src/main/kotlin/net/corda/node/services/keys/E2ETestKeyManagementService.kt
@@ -5,9 +5,9 @@ import net.corda.core.identity.PartyAndCertificate
 import net.corda.core.internal.ThreadBox
 import net.corda.core.node.services.IdentityService
 import net.corda.core.serialization.SingletonSerializeAsToken
-import net.corda.core.crypto.AliasPrivateKey
-import net.corda.node.internal.cryptoservice.BCCryptoService
-import net.corda.node.internal.cryptoservice.CryptoService
+import net.corda.core.crypto.internal.AliasPrivateKey
+import net.corda.node.services.keys.cryptoservice.BCCryptoService
+import net.corda.nodeapi.internal.cryptoservice.CryptoService
 import org.bouncycastle.operator.ContentSigner
 import java.security.KeyPair
 import java.security.PrivateKey
